refactor(gotjunk): Extract Liberty into LibertyClassificationContent component (Sprint 2/3)
Sprint 2/3 completed: Liberty-only logic extracted into dedicated component.

Changes:
1. New component: LibertyClassificationContent.tsx (417 lines)
   - Owns all Liberty-only state (stayLimitSheetOpen, alertTimerSheetOpen, etc.)
   - Owns all Liberty-only handlers (couchLongPress, campingLongPress, iceLongPress, policeLongPress)
   - Receives renderModalShell via props for consistent UI
   - Receives discountLongPress/bidLongPress for GotJunk categories in accordion

2. ClassificationModal.tsx refactored (863 → 419 lines, -51%)
   - Delegates Liberty branch to LibertyClassificationContent
   - Keeps only GotJunk state/handlers
   - Preserves shared renderModalShell and renderGotJunkList functions

3. New test: LibertyClassificationContent.test.tsx
   - Unit tests for Liberty component isolation

Benefits:
- Performance: Liberty hooks only initialize when Liberty mode enabled (100% overhead elimination)
- Maintainability: Clean component boundaries, easier testing
- WSP 87: ClassificationModal now 419 lines (well below 800 guideline)
- Code separation: GotJunk vs Liberty paths clearly isolated

Build: ✅ (839.41 kB, same as before - no bundle size increase)

WSP Compliance:
- WSP 84 (Code Memory): No duplication, proper reuse via props
- WSP 87 (File Size): 419 lines (improved from 863)
- WSP 50 (Reuse Existing): renderModalShell shared between components

Audit Reports:
- ClassificationModal_AUDIT_FINAL.md: Pre-extraction audit
- ClassificationModal_AUDIT_V2.md: Vibecoding detection
- ClassificationModal_DEEP_AUDIT.md: Deep analysis
