+++ modules/foundups/gotjunk/frontend/components/ClassificationModal_AUDIT.md
@@ -1,0 +1,232 @@
+# ClassificationModal.tsx - Vibecoding & Dead Code Audit
+
+**File**: `modules/foundups/gotjunk/frontend/components/ClassificationModal.tsx`  
+**Date**: Current  
+**Lines**: ~990  
+**Status**: ⚠️ VIBECODING DETECTED + DEAD CODE FOUND
+
+---
+
+## 🔴 CRITICAL VIBECODING ISSUES
+
+### 1. **Massive Code Duplication (3 Render Paths)**
+**Severity**: HIGH  
+**Lines**: 217-476, 482-639, 645-988
+
+**Problem**: The same modal structure is duplicated 3 times:
+- Map View (lines 217-476)
+- Regular Camera (lines 482-639)  
+- Liberty Camera (lines 645-988)
+
+**Duplicated Elements**:
+- Modal wrapper with same styles (lines 221-231, 486-496, 648-658)
+- Image preview logic (lines 233-242, 498-507, 660-669)
+- Title rendering (lines 244-247, 509-512, 671-674)
+- Cancel button (lines 612-620, 962-970)
+- Action Sheets (lines 447-471, 622-634, 972-984)
+
+**Impact**: 
+- ~300 lines of duplicated code
+- Maintenance nightmare - changes must be made in 3 places
+- Inconsistent styling/behavior risk
+- Violates DRY principle
+
+**Recommendation**: Extract common modal wrapper, image preview, and action sheets into reusable components or shared render functions.
+
+---
+
+### 2. **Unnecessary useLongPress Hook Usage**
+**Severity**: MEDIUM  
+**Lines**: 152-180
+
+**Problem**: Food subcategory handlers use `useLongPress` hook with empty `onLongPress` functions:
+
+```typescript
+const soupKitchenLongPress = useLongPress({
+  onLongPress: () => {}, // Empty function - unnecessary hook
+  onTap: () => onClassify('soup_kitchen'),
+  threshold: 450,
+});
+```
+
+**Affected Handlers**:
+- `soupKitchenLongPress` (line 152)
+- `bbqLongPress` (line 158)
+- `dryFoodLongPress` (line 164)
+- `pickLongPress` (line 170)
+- `gardenLongPress` (line 176)
+
+**Impact**:
+- Unnecessary hook overhead (5 instances)
+- Confusing code intent (why use long press if it does nothing?)
+- Performance overhead from hook initialization
+
+**Recommendation**: Replace with simple `onClick` handlers:
+```typescript
+onClick={() => onClassify('soup_kitchen')}
+```
+
+---
+
+### 3. **Unused State Setters**
+**Severity**: LOW  
+**Lines**: 83, 87
+
+**Problem**: State setters are declared but never used:
+
+```typescript
+const [discountPercent, setDiscountPercent] = useState(() => {
+  const saved = localStorage.getItem('gotjunk_default_discount');
+  return saved ? parseInt(saved, 10) : 75;
+});
+const [bidDurationHours, setBidDurationHours] = useState(() => {
+  const saved = localStorage.getItem('gotjunk_default_bid_duration');
+  return saved ? parseInt(saved, 10) : 72;
+});
+```
+
+**Impact**:
+- `setDiscountPercent` - Never called (values updated via localStorage in handlers)
+- `setBidDurationHours` - Never called (values updated via localStorage in handlers)
+- Unnecessary state management overhead
+
+**Recommendation**: Convert to simple constants or use `useMemo`:
+```typescript
+const discountPercent = useMemo(() => {
+  const saved = localStorage.getItem('gotjunk_default_discount');
+  return saved ? parseInt(saved, 10) : 75;
+}, []); // Re-read on mount only
+```
+
+---
+
+## 🟡 MODERATE ISSUES
+
+### 4. **Inconsistent Button Patterns**
+**Severity**: MEDIUM  
+**Pattern**: Mix of `motion.button` and `button` elements
+
+**Problem**: 
+- Some buttons use `motion.button` with animations (lines 257, 308, 324, etc.)
+- Others use plain `button` with `{...longPress}` spread (lines 273, 288, etc.)
+- No clear pattern for when to use which
+
+**Impact**: Inconsistent UX, harder to maintain
+
+**Recommendation**: Standardize on `motion.button` for all interactive elements, or create a `CategoryButton` component.
+
+---
+
+### 5. **Outdated Documentation Comments**
+**Severity**: LOW  
+**Lines**: 47-61
+
+**Problem**: Comment says "LIBERTY CAMERA (6 options)" but structure has changed:
+- Now uses accordion with 4 categories (Stuff, Alert, Food, Shelter)
+- Stuff category has 5 subcategories
+- Food has 5 subcategories
+- Alert has 2 subcategories  
+- Shelter has 3 subcategories
+
+**Impact**: Misleading documentation
+
+**Recommendation**: Update comments to reflect current accordion structure.
+
+---
+
+### 6. **Repeated Action Sheet Rendering**
+**Severity**: LOW  
+**Lines**: 447-471, 622-634, 972-984
+
+**Problem**: Same 4 Action Sheets rendered in all 3 render paths:
+- `ActionSheetDiscount`
+- `ActionSheetBid`
+- `ActionSheetStayLimit`
+- `ActionSheetAlertTimer`
+
+**Impact**: Unnecessary duplication, but acceptable if needed for conditional rendering
+
+**Recommendation**: Extract to shared function or component if all paths need them.
+
+---
+
+## 🟢 MINOR ISSUES
+
+### 7. **Magic Numbers**
+**Severity**: LOW  
+**Examples**: `threshold: 450`, `scale: 1.03`, `scale: 0.97`
+
+**Recommendation**: Extract to constants:
+```typescript
+const LONG_PRESS_THRESHOLD = 450;
+const HOVER_SCALE = 1.03;
+const TAP_SCALE = 0.97;
+```
+
+---
+
+### 8. **Hardcoded Default Values**
+**Severity**: LOW  
+**Lines**: 85, 89
+
+**Problem**: Default values (75, 72) hardcoded in multiple places
+
+**Recommendation**: Extract to constants:
+```typescript
+const DEFAULT_DISCOUNT_PERCENT = 75;
+const DEFAULT_BID_DURATION_HOURS = 72;
+```
+
+---
+
+## 📊 SUMMARY
+
+### Dead Code Found:
+- ✅ `setDiscountPercent` - Never used
+- ✅ `setBidDurationHours` - Never used  
+- ✅ 5 food subcategory `useLongPress` hooks with empty handlers
+
+### Vibecoding Patterns:
+- 🔴 **CRITICAL**: 3 duplicate render paths (~300 lines duplicated)
+- 🟡 **MODERATE**: Unnecessary hook usage (5 instances)
+- 🟡 **MODERATE**: Inconsistent button patterns
+- 🟢 **MINOR**: Magic numbers, hardcoded values
+
+### Estimated Cleanup Impact:
+- **Lines Removed**: ~350-400 (after deduplication)
+- **Maintainability**: ⬆️ Significantly improved
+- **Performance**: ⬆️ Slight improvement (fewer hooks, less code)
+- **Risk**: 🟡 Medium (requires careful refactoring)
+
+---
+
+## 🎯 RECOMMENDED ACTIONS
+
+### Priority 1 (High Impact):
+1. **Extract common modal wrapper** - Create `<ModalWrapper>` component
+2. **Extract image preview** - Create `<ImagePreview>` component  
+3. **Remove unnecessary useLongPress hooks** - Replace with onClick
+
+### Priority 2 (Medium Impact):
+4. **Convert unused state setters** - Use constants or useMemo
+5. **Standardize button patterns** - Create `<CategoryButton>` component
+6. **Extract action sheets** - Shared render function
+
+### Priority 3 (Low Impact):
+7. **Extract magic numbers** - Constants file
+8. **Update documentation** - Reflect current structure
+
+---
+
+## ✅ WSP COMPLIANCE
+
+**WSP 84 (Code Memory Verification)**: ⚠️ VIOLATION
+- Code duplication indicates vibecoding (recreating instead of reusing)
+
+**WSP 87 (File Size)**: ✅ COMPLIANT
+- Current: ~990 lines
+- Guideline: 800-1000
+- Hard limit: 1500
+
+**Recommendation**: Address Priority 1 items to restore WSP 84 compliance.
+
