+++ modules/foundups/gotjunk/frontend/components/ClassificationModal_AUDIT_V2.md
@@ -1,0 +1,206 @@
+# ClassificationModal.tsx - Vibecoding & Dead Code Audit V2
+
+**File**: `modules/foundups/gotjunk/frontend/components/ClassificationModal.tsx`  
+**Date**: Current  
+**Lines**: ~824  
+**Status**: ⚠️ VIBECODING DETECTED + DEAD CODE FOUND
+
+---
+
+## 🔴 CRITICAL VIBECODING ISSUES
+
+### 1. **Code Duplication: Map View vs Regular Camera**
+
+**Severity**: HIGH  
+**Lines**: 251-365 (Map View), 372-501 (Regular Camera)
+
+**Problem**: Both render paths show identical GotJunk categories (Free, Discount, Bid, Share, Wanted) with nearly identical structure:
+
+**Map View** (lines 251-365):
+- Same 5 categories
+- Same button structure
+- Same action sheets (Discount, Bid)
+- Only difference: Styling (no icons, simpler layout)
+
+**Regular Camera** (lines 372-501):
+- Same 5 categories
+- Same button structure  
+- Same action sheets (Discount, Bid)
+- Only difference: Icons (FreeIcon, DiscountIcon, BidIcon), shadow effects
+
+**Impact**: 
+- ~100 lines of duplicated category rendering
+- Changes must be made in 2 places
+- Inconsistent styling between views
+
+**Recommendation**: Extract GotJunk category buttons into shared component or function.
+
+---
+
+### 2. **Liberty-Only Handlers Always Initialized**
+
+**Severity**: MEDIUM  
+**Lines**: 112-146
+
+**Problem**: Liberty-only handlers are always created, even when Liberty is disabled:
+
+```typescript
+const couchLongPress = useLongPress({...});      // Only used when libertyEnabled
+const campingLongPress = useLongPress({...});    // Only used when libertyEnabled
+const iceLongPress = useLongPress({...});        // Only used when libertyEnabled
+const policeLongPress = useLongPress({...});     // Only used when libertyEnabled
+```
+
+**Impact**:
+- Unnecessary hook initialization when Liberty is OFF
+- Performance overhead (4 hooks initialized but unused)
+- Memory overhead
+
+**Recommendation**: Conditionally initialize Liberty handlers only when `libertyEnabled === true`.
+
+---
+
+### 3. **Liberty-Only State Always Initialized**
+
+**Severity**: MEDIUM  
+**Lines**: 74-77, 80
+
+**Problem**: Liberty-only state is always initialized, even when Liberty is disabled:
+
+```typescript
+const [stayLimitSheetOpen, setStayLimitSheetOpen] = useState(false);
+const [alertTimerSheetOpen, setAlertTimerSheetOpen] = useState(false);
+const [currentStayLimitType, setCurrentStayLimitType] = useState<'couch' | 'camping'>('couch');
+const [currentAlertType, setCurrentAlertType] = useState<'ice' | 'police'>('police');
+const [expandedCategory, setExpandedCategory] = useState<'stuff' | 'alert' | 'food' | 'shelter' | null>(null);
+```
+
+**Impact**:
+- Unnecessary state initialization when Liberty is OFF
+- Memory overhead (5 state variables unused)
+
+**Recommendation**: Conditionally initialize Liberty state only when `libertyEnabled === true`, or use lazy initialization.
+
+---
+
+## 🟡 MODERATE ISSUES
+
+### 4. **Outdated Interface Comment**
+
+**Severity**: LOW  
+**Line**: 34
+
+**Problem**: Comment says "Show all 11 categories on map" but Map View now shows only 5 GotJunk categories:
+
+```typescript
+isMapView?: boolean; // Show all 11 categories on map, only 5 on My Items
+```
+
+**Reality**: Map View shows 5 categories (same as Regular Camera) when Liberty is OFF.
+
+**Recommendation**: Update comment to reflect current behavior:
+```typescript
+isMapView?: boolean; // Show GotJunk categories (5 types) on map when Liberty is OFF
+```
+
+---
+
+### 5. **Outdated Documentation Comments**
+
+**Severity**: LOW  
+**Lines**: 47-61
+
+**Problem**: Comment says "LIBERTY CAMERA (6 options)" but structure has changed:
+- Now uses accordion with 4 categories (Stuff, Alert, Food, Shelter)
+- Stuff category has 5 subcategories
+- Food has 5 subcategories
+- Alert has 2 subcategories
+- Shelter has 3 subcategories
+- Total: 15+ options, not 6
+
+**Recommendation**: Update documentation to reflect accordion structure.
+
+---
+
+### 6. **Unused Imports (Conditional)**
+
+**Severity**: LOW  
+**Lines**: 26-27
+
+**Problem**: `ActionSheetStayLimit` and `ActionSheetAlertTimer` are imported but only used when Liberty is enabled:
+
+```typescript
+import { ActionSheetStayLimit } from './ActionSheetStayLimit';
+import { ActionSheetAlertTimer } from './ActionSheetAlertTimer';
+```
+
+**Impact**: 
+- Unnecessary bundle size when Liberty is disabled
+- Minor - imports are tree-shakeable in modern bundlers
+
+**Recommendation**: Keep as-is (acceptable for code clarity), or use dynamic imports if bundle size becomes an issue.
+
+---
+
+## 🟢 MINOR ISSUES
+
+### 7. **Inconsistent Button Patterns**
+
+**Severity**: LOW  
+**Pattern**: Mix of `motion.button` and `button` elements
+
+**Problem**: 
+- Simple buttons (Free, Share, Wanted, Housing) use `motion.button`
+- Buttons with long press (Discount, Bid, Couch, Camping, ICE, Police) use `button` with `{...longPress}` spread
+
+**Impact**: Inconsistent UX patterns, but acceptable for functionality
+
+**Recommendation**: Standardize if UX consistency becomes priority.
+
+---
+
+## 📊 SUMMARY
+
+### Dead Code Found:
+- ✅ None (all code is used, but conditionally)
+
+### Vibecoding Patterns:
+- 🔴 **CRITICAL**: Map View and Regular Camera duplicate GotJunk category rendering (~100 lines)
+- 🟡 **MODERATE**: Liberty handlers/state always initialized (performance overhead)
+- 🟡 **MODERATE**: Outdated comments/documentation
+
+### Estimated Cleanup Impact:
+- **Lines Removed**: ~100 (after deduplicating GotJunk categories)
+- **Performance**: ⬆️ Improved (conditional Liberty initialization)
+- **Maintainability**: ⬆️ Significantly improved
+- **Risk**: 🟡 Medium (requires careful refactoring)
+
+---
+
+## 🎯 RECOMMENDED ACTIONS
+
+### Priority 1 (High Impact):
+1. **Extract GotJunk category buttons** - Create shared component/function for Free, Discount, Bid, Share, Wanted
+2. **Conditionally initialize Liberty handlers** - Only create when `libertyEnabled === true`
+
+### Priority 2 (Medium Impact):
+3. **Conditionally initialize Liberty state** - Use lazy initialization or conditional hooks
+4. **Update outdated comments** - Fix interface and documentation comments
+
+### Priority 3 (Low Impact):
+5. **Standardize button patterns** - If UX consistency becomes priority
+
+---
+
+## ✅ WSP COMPLIANCE
+
+**WSP 84 (Code Memory Verification)**: ⚠️ VIOLATION
+- Code duplication indicates vibecoding (recreating instead of reusing)
+
+**WSP 87 (File Size)**: ✅ COMPLIANT
+- Current: ~824 lines
+- Guideline: 800-1000
+- Hard limit: 1500
+
+**Recommendation**: Address Priority 1 items to restore WSP 84 compliance.
+
